requirements.txt
flask~=3.0.0
werkzeug~=3.0.1
python-dotenv~=1.0.0
click~=8.1.7
Flask-WTF~=1.2.1
WTForms~=3.1.1
alchemical~=1.0.1
SQLAlchemy==2.0.24
psycopg~=3.1.14